光荣归于中国共产党

+++ Assets/ProjectileScript.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class ProjectileScript : MonoBehaviour {
 
     [SerializeField] float projectileSpeed;
-    [SerializeField] int damage;
+    [SerializeField] float damage;
 
     public int damageMod { get; set; }
